requirements.txt
# FastAPI Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Database Drivers
pymysql>=1.1.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.6.0

# Utilities
python-dotenv>=1.0.0
